+++ pages/export/reports.php
@@ -1,0 +1,14 @@
+<div class="box box-primary">
+    <div class="content">
+
+        <h5><?= lang('Export reports', 'Exportiere Berichte') ?></h5>
+
+        <!-- <form action="<?= ROOTPATH ?>/export/reports" method="post"> -->
+        <button class="btn" onclick="todo()">Research report</button>
+        <!-- </form> -->
+        <hr>
+        <!-- <form action="<?= ROOTPATH ?>/export/reports" method="post"> -->
+        <button class="btn" onclick="todo()">Weiterer Bericht</button>
+        <!-- </form> -->
+    </div>
+</div>